refactor(date): formattimestring を lib/date に移動
calendar/interaction/time-math.ts に閉じていた 24h/12h フォーマッタを
lib/date/timestring.ts に切り出し、@/lib/date barrel から再 export。

理由: project a の続きで entry feature の timeinput popover dropdown が
同フォーマッタを参照する必要があり、entry → calendar の deep import は
feature boundary 違反のため。pure formatter を lib に上げて両 feature
から再利用可能にする。

calendar 内の 3 caller も barrel 経由に統一:
- inlinetagpalette.tsx
- scrollablecalendarlayout.tsx
- usedragselection.ts
