+++ src/features/calendar/interaction/time-math.ts
@@ -43,20 +43,6 @@ export function snapToGrid(
   return { snappedTop, hour, minute };
 }
 
-/** 時刻フォーマット（24h: "14:30" / 12h: "2:30 PM"） */
-export function formatTimeString(
-  hour: number,
-  minute: number,
-  format: '24h' | '12h' = '24h',
-): string {
-  if (format === '12h') {
-    const period = hour >= 12 ? 'PM' : 'AM';
-    const h12 = hour === 0 ? 12 : hour > 12 ? hour - 12 : hour;
-    return `${h12}:${String(minute).padStart(2, '0')} ${period}`;
-  }
-  return `${String(hour).padStart(2, '0')}:${String(minute).padStart(2, '0')}`;
-}
-
 /** 時刻 + duration → 終了時刻 */
 export function addMinutesToTime(
   hour: number,
